Refactor galaxy.jobs.error_level -> galaxy.tools.parser.error_level.
Needed after 46170e9bcdcc2eb104ebab72a7afc9b7238bc314 to keep galaxy-lib from importing things outside of galaxy-lib but clearly a good refactoring on its own since it eliminates mid-file imports that were there to prevent circular dependencies from loading Galaxy.
