+++ lib/galaxy/tools/parser/xml.py
@@ -8,6 +8,7 @@ from math import isinf
 from galaxy.tools.deps import requirements
 from galaxy.util import string_as_bool, xml_text, xml_to_string
 from galaxy.util.odict import odict
+from .error_level import StdioErrorLevel
 from .interface import (
     InputSource,
     PageSource,
@@ -861,7 +862,6 @@ class StdioParser(object):
         Parses error level and returns error level enumeration. If
         unparsable, returns 'fatal'
         """
-        from galaxy.jobs.error_level import StdioErrorLevel
         return_level = StdioErrorLevel.FATAL
         try:
             if err_level:
